add new abstractClass in net
